changed abstractions structure
